validate outputs in e2e test, cosmetics
