+++ tests/unittest/llmapi/lora_test_utils.py
@@ -1,5 +1,11 @@
+# NeMo LoRA test utilities
+import json
+import tarfile
+import tempfile
+from pathlib import Path
 from typing import OrderedDict, Type
 
+import torch
 from utils.llm_data import llm_models_root
 from utils.util import duplicate_list_to_length, flatten_list, similar
 
@@ -114,3 +120,95 @@ def check_llama_7b_multi_lora_from_request_test_harness(
     for output, ref, key_word in zip(outputs, references, key_words):
         assert similar(output.outputs[0].text,
                        ref) or key_word in output.outputs[0].text
+
+
+def create_mock_nemo_lora_checkpoint(
+    lora_dir: Path,
+    hidden_size: int = 4096,
+    num_layers: int = 32,
+    lora_rank: int = 8,
+    tp_size: int = 1,
+    num_attention_heads: int = 32,
+    num_kv_heads: int = None,  # If None, defaults to num_attention_heads
+    dtype: torch.dtype = torch.float16,
+) -> Path:
+    """Create a minimal NeMo LoRA checkpoint for testing.
+
+    This creates a .nemo tarfile with the expected structure:
+    - model_weights.ckpt containing attn_qkv adapter weights
+    - model_config.yaml with basic configuration
+
+    Args:
+        lora_dir: Directory to create the checkpoint in
+        hidden_size: Model hidden size
+        num_layers: Number of transformer layers
+        lora_rank: LoRA rank
+        tp_size: Tensor parallelism size
+        num_attention_heads: Number of query attention heads
+        num_kv_heads: Number of key/value heads (for GQA). If None, equals num_attention_heads
+        dtype: Data type for the weights (default: torch.float16)
+
+    Returns:
+        Path to the created .nemo file
+    """
+    # Default to standard MHA if not specified
+    if num_kv_heads is None:
+        num_kv_heads = num_attention_heads
+
+    nemo_path = lora_dir / "test_lora.nemo"
+
+    with tempfile.TemporaryDirectory() as temp_dir_str:
+        temp_dir = Path(temp_dir_str)
+
+        weights_dict = {}
+
+        head_dim = hidden_size // num_attention_heads
+        kv_hidden_size = head_dim * num_kv_heads
+
+        qkv_output_dim = hidden_size + 2 * kv_hidden_size
+
+        for layer_idx in range(num_layers):
+            key_prefix = f"model.layers.{layer_idx}.self_attention.adapter_layer.lora_kqv_adapter"
+
+            # Create linear_in weights [lora_rank, hidden_size] with small random values
+            linear_in_key = f"{key_prefix}.linear_in.weight"
+            weights_dict[linear_in_key] = torch.randn(
+                lora_rank, hidden_size, dtype=dtype) * 0.01
+
+            # Create linear_out weights [qkv_output_dim, lora_rank] for fused QKV
+            # This is the key difference for GQA - the output dimension changes
+            linear_out_key = f"{key_prefix}.linear_out.weight"
+            weights_dict[linear_out_key] = torch.randn(
+                qkv_output_dim, lora_rank, dtype=dtype) * 0.01
+
+        ckpt_path = temp_dir / "model_weights.ckpt"
+        torch.save(weights_dict, ckpt_path)
+
+        config = {
+            "precision": "fp16" if dtype == torch.float16 else "bf16",
+            "trainer": {
+                "num_nodes": 1,
+                "devices": tp_size,
+            },
+            "model": {
+                "hidden_size": hidden_size,
+                "num_layers": num_layers,
+                "num_attention_heads": num_attention_heads,
+                "num_query_groups": num_kv_heads,  # This is the key for GQA
+            },
+            "lora": {
+                "rank": lora_rank,
+                "target_modules": ["attn_qkv"],
+            }
+        }
+
+        config_path = temp_dir / "model_config.yaml"
+        # Using JSON for simplicity since YAML parsing isn't critical for the test
+        with open(config_path, 'w') as f:
+            json.dump(config, f)
+
+        with tarfile.open(nemo_path, 'w') as tar:
+            tar.add(ckpt_path, arcname="model_weights.ckpt")
+            tar.add(config_path, arcname="model_config.yaml")
+
+    return nemo_path
